Rename enforcement action search to filter
Renaming the search page to filter page to more closely match what the
page is actually called. The page refers to the input as filters and is
for applying filters, not searching. The distinction is subtle, but
since the default page gives us the complete list of enforcement actions
to begin with, we are filtering. If we started with a blank slate, then
we'd be searching.