Add scraping for enforcement action detail page
The url for an enforcement action's detail page depends on the title of
the enforcement action. The detail page provides sumamry info related to the
enforcement action such as brief description, docket number, and date.
It also links to the associated pres release statement, which provides
more detailed information we desire.