Make the url builder functions idempotent
Instead of concatenating straight to the url on each parameter function
call, adjust the functions so that each call sets the parameter. Move
the string concatenation to the build function.

This makes it so the order in which the parameters are set doesn't
matter. This also enseures no accidental adding of a parameteter more
than needed (e.g. setting 2 different from dates).