Add parser check for last page
Since the results of our filtered enforcement actions are paginated, we
need to iterate through all the pages extracting the detail links. In
order to know when we've reached the last page to iterate, we need to
check if the next page button is enabled or disabled. When it's
disabled, we can assume we the current page is the last one.